requirements.txt
python-dotenv==1.0.1
requests==2.32.3
selenium==4.26.1
twilio==9.3.7